Feature/order

+++ src/main/java/com/example/onlinemarketbe/model/Order.java
@@ -5,22 +5,30 @@ import lombok.Setter;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
-import java.sql.Date;
+import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
+
+import java.util.HashSet;
+import java.util.Set;
+
 
 
 @Getter
 @Setter
 @Entity
+
+
 @Table(name = "orders")
+
+
 public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     @Column
     @NotNull
-    private int totalPrice;
+    private double totalPrice;
     @ManyToOne
     @JoinColumn(name = "payment_id")
     Payment payment;
@@ -33,11 +41,10 @@ public class Order {
     @ManyToOne
     @JoinColumn(name = "status_id")
     Status status;
-
-
     @OneToMany(fetch = FetchType.EAGER)
 //    @JoinTable( name = "order_item",
 //            joinColumns = @JoinColumn(name = "order_id"),
 //            inverseJoinColumns = @JoinColumn(name = "item_id"))
     private Set<Item> items = new HashSet<>();
+
 }
